Updated testing based on review.
